refactored for modular script adaptation
