+++ cli_FSD/legacy/__init__.py
@@ -1,0 +1,20 @@
+# __init__.py for cli_FSD package
+from .main import main
+from .main import clean_up_llm_response
+from .main import create_bash_invocation_script
+from .main import consult_llm_for_error_resolution
+from .main import chat_with_model
+from .main import cleanup_previous_assembled_scripts
+from .main import execute_shell_command
+from .main import  cleanup_previous_assembled_scripts
+from .main import execute_script_with_repl_and_consultation
+from .main import extract_script_from_response
+
+from .resources.assembler import AssemblyAssist
+
+
+
+# You can import specific classes or functions from your modules here
+# to make them available at the package level, e.g.,
+# from .main import main_function
+
