--- cli_FSD/v2/__init__.py
@@ -1,1 +1,0 @@
-from .main import main
